Fix: Do not check if wireless-capable device is installed
I am pretty sure that this will cause errors in the future since it only checks if the radio is enabled. There doesn't seem to be an easy method that covers all distros and devices. Since this is not necessary for the script to finish correctly, we can scrap this.